+++ client/packages/executor/.env-tests
@@ -1,0 +1,1 @@
+ETHEREUM_PRIVATE_KEY=0x9ece8cb36ec09fa95ad430c77b5c00c6a2ffb080e2bfb5074383ab2658f1145f
